% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulY.R
\name{cumulY}
\alias{cumulY}
\title{Function to compute the cumulative prediction value of longitudinal outcome using an history window}
\usage{
cumulY(predRE, data, time, marker_name, tLM, HW)
}
\arguments{
\item{HW}{}
}
\description{
Function to compute the cumulative prediction value of longitudinal outcome using an history window
}
\examples{

}
