% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivY.R
\name{derivY}
\alias{derivY}
\title{Function to predict the slope of longitudinal outcome}
\usage{
derivY(predRE, data, derivForm, time, tLM)
}
\arguments{
\item{predRE}{A list object from \code{predRE} function}

\item{data}{A dataframe where each row containing some predictive variables for a specific subject}

\item{derivForm}{A list containing the derivation form}
}
\value{
A matrix containing the prediction slope of the longitudinal outcome for each subject
}
\description{
Function to predict the slope of longitudinal outcome
}
\examples{

}
