% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark.R
\name{landmark}
\alias{landmark}
\title{Function to compute several summaries of longitudinal outcomes}
\usage{
landmark(lmm_objects, data, tLM, subject, time, derivForm_objects)
}
\arguments{
\item{lmm_objects}{A list containing as many \code{hlme} output as longitudinal outcome}

\item{data}{A dataframe containing longitudinal data}

\item{tLM}{An integer indicating the landmark time}

\item{subject}{A character indicating the subject variable}

\item{time}{A character indicating the time-dependent variable}

\item{derivForm_objects}{A list containing as many derivation form as longitudinal outcome}
}
\value{
The original dataframe including the summaries compute from longitudinal outcomes
}
\description{
Function to compute several summaries of longitudinal outcomes
}
\examples{

}
