% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMpred.R
\name{LMpred}
\alias{LMpred}
\title{Title}
\usage{
LMpred(method = c("LM-cox", "LM-rsf", "cox", "LM-coxnet", "LM-splsDR"),
  models, newdata, time, subject, marker_list = NULL,
  covar_list = NULL, tHor)
}
\arguments{
\item{tHor}{}
}
\description{
Title
}
\examples{

}
