% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsum_mfpc.R
\name{LMsum.mfpc}
\alias{LMsum.mfpc}
\title{Title}
\usage{
LMsum.mfpc(data, data.surv, data.pred, data.surv.pred, markers, tLM,
  subject, time, time.interval = 0.1, pve = 0.95, nbasis = 3)
}
\arguments{
\item{nbasis}{}
}
\description{
Title
}
