% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsum.R
\name{LMsum}
\alias{LMsum}
\title{Function to compute several summaries of longitudinal outcomes}
\usage{
LMsum(
  lmm_objects,
  data,
  tLM,
  subject,
  time,
  derivForm_objects,
  HW = tLM,
  threshold = NULL
)
}
\arguments{
\item{lmm_objects}{A list containing as many \code{hlme} output as longitudinal outcome}

\item{data}{A dataframe containing longitudinal data}

\item{tLM}{An integer indicating the landmark time}

\item{subject}{A character indicating the subject variable}

\item{time}{A character indicating the time-dependent variable}

\item{derivForm_objects}{A list containing as many derivation form as longitudinal outcome}

\item{HW}{An integer meaning the range of history window for computing the cumulative values.
Default is \code{HW = tLM} resulting of the entire history.}

\item{threshold}{A list containing a numeric threshold}
}
\value{
The original dataframe including the summaries compute from longitudinal outcomes
}
\description{
Function to compute several summaries of longitudinal outcomes
}
\examples{

}
