% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsum_simu.R
\name{LMsum_simu}
\alias{LMsum_simu}
\title{Internal function for simulations}
\usage{
LMsum_simu(markers.true, data, tLM, subject, time, RE, HW = tLM,
  threshold = NULL)
}
\description{
Internal function for simulations
}
\examples{

}
