% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsurv_penalized_cox.R
\name{LMsurv.coxnet}
\alias{LMsurv.coxnet}
\title{Title}
\usage{
LMsurv.coxnet(data.surv, coxnet.opt, coxnet.lasso, coxnet.ridge)
}
\arguments{
\item{coxnet.ridge}{}
}
\value{

}
\description{
Title
}
