% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsurv_rsf.R
\name{LMsurv.rsf}
\alias{LMsurv.rsf}
\title{Title}
\usage{
LMsurv.rsf(
  data.surv,
  rsf.split,
  rsf.submodels,
  cause = 1,
  CR = FALSE,
  nodesize.grid = NULL,
  mtry.grid = NULL
)
}
\arguments{
\item{CR}{}
}
\value{

}
\description{
Title
}
