% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsurv_spls.R
\name{LMsurv.spls}
\alias{LMsurv.spls}
\title{Title}
\usage{
LMsurv.spls(
  data.surv,
  spls.opt = TRUE,
  spls.nosparse = !spls.opt,
  spls.maxsparse = !spls.opt
)
}
\arguments{
\item{spls.maxsparse}{}
}
\value{

}
\description{
Title
}
