% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_LMpred.R
\name{old_LMpred}
\alias{old_LMpred}
\title{Title}
\usage{
old_LMpred(
  method = c("LM-cox", "LM-rsf", "cox", "LM-coxnet", "LM-splsDR"),
  models,
  newdata,
  time,
  subject,
  marker_list = NULL,
  covar_list = NULL,
  tHor
)
}
\arguments{
\item{tHor}{}
}
\value{

}
\description{
Title
}
\examples{

}
