% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRE_binary_simu.R
\name{predRE.binary_simu}
\alias{predRE.binary_simu}
\title{Internal function for simulations}
\usage{
predRE.binary_simu(model, data)
}
\description{
Internal function for simulations
}
\examples{

}
