% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRE_lcmm.R
\name{predRE.lcmm}
\alias{predRE.lcmm}
\title{Function to predict the random effects on data using lcmm output}
\usage{
predRE.lcmm(model, formul, data, subject)
}
\arguments{
\item{model}{A \code{lcmm} object}

\item{data}{A dataframe containing longitudinal data}

\item{subject}{}
}
\value{
A list containing the random effect for each subject, the fixed coefficients
from \code{lcmm} and the formula used to compute the \code{lcmm} mixed model
}
\description{
Function to predict the random effects on data using lcmm output
}
\examples{

}
