% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRE_lme4.R
\name{predRE.lme4}
\alias{predRE.lme4}
\title{Function to predict the random effects on data using lme4 output}
\usage{
predRE.lme4(model, data)
}
\arguments{
\item{model}{A \code{lme4} object}

\item{data}{A dataframe containing longitudinal data}
}
\value{
A list containing the random effect for each subject, the fixed coefficients
from \code{lme4} and the formula used to compute the \code{lme4} mixed model
}
\description{
Function to predict the random effects on data using lme4 output
}
\examples{

}
