% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRE.R
\name{predRE}
\alias{predRE}
\title{Function to predict the random effects on data using hlme output}
\usage{
predRE(model, data)
}
\arguments{
\item{model}{A \code{hlme} object}

\item{data}{A dataframe containing longitudinal data}
}
\value{
A list containing the random effect for each subject, the fixed coefficients
from \code{hlme} and the formula used to compute the \code{hlme} mixed model
}
\description{
Function to predict the random effects on data using hlme output
}
\examples{

}
