% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predY.R
\name{predY}
\alias{predY}
\title{Function to predict the value of longitudinal outcome for a specific time}
\usage{
predY(predRE, data, time, tLM)
}
\arguments{
\item{predRE}{A list object from \code{predRE} function}

\item{data}{A dataframe where each row containing some predictive variables for a specific subject}
}
\value{
A matrix containing the prediction value of the longitudinal outcome for each subject
}
\description{
Function to predict the value of longitudinal outcome for a specific time
}
\examples{

}
