% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdlandmark.R
\name{hdlandmark}
\alias{hdlandmark}
\title{Title}
\usage{
hdlandmark(data, data.pred = NULL, markers, tLMs, tHors, subject, time,
  time.event, event, long.methods = c("combine", "both", "GLMM", "MFPC"),
  surv.methods = c("all", "cox", "penalized-cox", "sPLS", "RSF"))
}
\arguments{
\item{surv.methods}{}
}
\description{
Title
}
