% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdlandmark.R
\name{hdlandmark}
\alias{hdlandmark}
\title{Title}
\usage{
hdlandmark(data, data.pred = NULL, markers, tLMs, tHors, subject, time,
  time.event, event, long.method = c("combine", "GLMM", "MFPC"),
  lmm.package = c("lme4", "lcmm"), cox.submodels = c("autoVar",
  "allVar"), coxnet.submodels = c("opt", "lasso", "ridge"),
  spls.submodels = c("opt", "nosparse", "maxsparse"),
  rsf.submodels = c("opt", "noVS", "default"), rsf.split = c("logrank",
  "bs.gradient"), kfolds = 10, seed = 1234, parallel = FALSE,
  Ncpus = 1)
}
\arguments{
\item{Ncpus}{}
}
\description{
Title
}
