% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMpred.R
\name{LMpred}
\alias{LMpred}
\title{Title}
\usage{
LMpred(
  data.surv,
  model.surv,
  long.method,
  surv.methods,
  tHor,
  cause = 1,
  CR = FALSE
)
}
\arguments{
\item{CR}{}
}
\value{

}
\description{
Title
}
