% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsummaries.R
\name{LMsummaries}
\alias{LMsummaries}
\title{Title}
\usage{
LMsummaries(data, data.pred, markers, tLM, subject, time, time.event,
  event, long.method, lmm.package, surv.covar, scaling)
}
\arguments{
\item{scaling}{}
}
\description{
Title
}
