% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsurv.R
\name{LMsurv}
\alias{LMsurv}
\title{Title}
\usage{
LMsurv(
  data.surv,
  surv.methods,
  cox.submodels,
  coxnet.submodels,
  spls.submodels,
  rsf.submodels,
  rsf.split
)
}
\arguments{
\item{rsf.split}{}
}
\value{

}
\description{
Title
}
