% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsurv.R
\name{LMsurv}
\alias{LMsurv}
\title{Title}
\usage{
LMsurv(method = c("LM-cox", "LM-rsf", "cox", "LM-coxnet", "LM-splsDR"),
  models, newdata, time, subject, tHor)
}
\arguments{
\item{tHor}{}
}
\description{
Title
}
\examples{

}
