% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsurv_rsf.R
\name{LMsurv.rsf}
\alias{LMsurv.rsf}
\title{Title}
\usage{
LMsurv.rsf(
  data.surv,
  rsf.split = c("logrank", "bs.gradient"),
  rsf.opt = TRUE,
  rsf.noVS = !rsf.opt,
  rsf.default = !rsf.opt
)
}
\arguments{
\item{rsf.default}{}
}
\value{

}
\description{
Title
}
